GT3 Radial Transport Branch 2/27 commit
Date: 2/27/2019
Time: 12:09pm

Including debug functions in radial transport to verify source terms, fluxes, etc. are reasonable. Can be triggered by passing debugFlag=True to radialtransport. The neutrals calculation and related calculations (e.g., Sion) can also be deactivated by passing neutFlag=False to radialtransport.

gt3.py now includes modules that let you interactively create input files from commandline. The input files are saved to the input folder as togt3_shotnum_timeid. See the code comments. The main entrance to the interactive mode is gt3Prep. You should provide the shot id, timeid, and run id, as well as the filename. The "quiet" parameter disables interactive file generation and tells the code to look for an input file. If no input file is found, program exits. This will be helpful when/if we want to run gt3 on many shots at once and we want the code to not bug us.